% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMf.R
\name{LMf}
\alias{LMf}
\title{Landmarking Model Fitting Function}
\usage{
LMf (data,tw=list(sl=sl,w=w),cov=list(fixed=fixed,vary=vary),id=id,rtime=rtime,time=time,status=status,inter=T,func_covars=c("linear","quadratic"),func_lms=c("linear","quadratic"))
}
\arguments{
\item{data}{A data.frame containing the longitudinal and survival data.}

\item{tw}{A list containing two elements: `sl` (landmark time points) and `w` (prediction window).}

\item{cov}{A list containing two elements: `fixed` (fixed covariates) and `vary` (time-varying covariates).}

\item{id}{Character string specifying the subject ID variable}

\item{rtime}{Observation time column name}

\item{time}{Character string specifying the event time variable.}

\item{status}{Character string specifying the event status variable.}

\item{pre_cov}{Character vector of pre-specified covariates with dynamic effects(optional)}

\item{inter}{Logical indicating whether to include interactions with landmark times (default=TRUE)}

\item{func_covars}{The functional form of the dynamic effect of covariates. When inter is TRUE, the available functional forms include linear functions (linear), quadratic functions (quadratic), exponential functions (exp), and logarithmic functions (log), with the default being linear and quadratic functions.}

\item{func_lms}{The functional forms for the dynamic effects of baseline hazard rates. When inter is TRUE, the available functional forms include linear functions (linear), quadratic functions (quadratic), exponential functions (exp), and logarithmic functions (log), with the default being linear and quadratic functions.}
}
\value{
An object of class "LMf" containing:
 \describe{
   \item{Model}{The result of parameter estimation.}
   \item{data}{The super landmark dataset.}
   \item{func_covars}{Functional forms for the dynamic effects of covariates.}
   \item{func_lms}{Functional forms for the dynamic effects of baseline hazard rates.}
   \item{cov}{A list of covariates type. The components of this list are:
     \itemize{
       \item{fixed}{Baseline covariates.}
       \item{vary}{Longitudinal covariates.}
     }
   }
   \item{tw}{A list of time points. The components of this list are:
     \itemize{
       \item{sl}{Landmark time points.}
       \item{w}{Prediction window.}
     }
   }
   \item{time}{Event time variable name.}
   \item{status}{Event status variable name.}
   \item{id}{Subject ID variable name.}
   \item{rtime}{Observation time column name (if time-varying covariates present).}
 }
}
\description{
This function fits landmark models for dynamic prediction using Cox proportional hazards models.
It handles both fixed and time-varying covariates and can automatically generate interaction terms
between covariates and landmark times.
}
\examples{
library(DynamicModel)
data(renal)
head(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
print(fit)
}
\references{
Fries AH, Choi E, Wu JT, Lee JH, Ding VY, Huang RJ, et al. Software Application Profile: dynamicLM —a tool for performing dynamic risk prediction using a landmark supermodel for survival data under competing risks. International Journal of Epidemiology 2023;52:1984–9. https://doi.org/10.1093/ije/dyad122.

Yang Z, Hou Y, Lyu J, Liu D, Chen Z. Dynamic prediction and prognostic analysis of patients with cervical cancer: a landmarking analysis approach. Annals of Epidemiology 2020;44:45–51. https://doi.org/10.1016/j.annepidem.2020.01.009.
}
