% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cHR.R
\name{plot_cHR}
\alias{plot_cHR}
\title{Title Function for dynamic effect hazard ratio (HR) plots.}
\usage{
plot_cHR(object,covars,conf_int=TRUE,IDlegend=NULL,legend=NULL,xlab="landmark time points",ylab="Dynamic HR",ylim,...)
}
\arguments{
\item{object}{An objects inheriting from class "LMf" or "VS_LM"}

\item{covars}{Covariate names.}

\item{conf_int}{Indicator variable. It indicates whether confidence intervals are included, with the default being TRUE.}

\item{IDlegend}{Label of the plot, with the default being NULL.}

\item{legend}{Annotations or descriptions of the plot, with the default being NULL.}

\item{xlab}{Label of the horizontal axis, with the default being "landmark time points(s)".}

\item{ylab}{Label of the vertical axis, with the default being "Dynamic HR (w=5)".}

\item{ylim}{The scale range of the vertical axis is, by default, from the minimum to the maximum value of effect size of the covariate.}

\item{...}{extra aguments; currently none is used.}
}
\description{
Title Function for dynamic effect hazard ratio (HR) plots.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
a1<-VS_LM(fit,cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")))
plot_cHR(a1,c("age"),conf_int=TRUE,IDlegend="a.",legend="(Per 10 years)",xlab="landmark time points(years)",ylab="cHR")
}
