% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamicHR.R
\name{plot_dynamicHR}
\alias{plot_dynamicHR}
\title{Title Dynamic Effects Plot of Covariates in a Joint Model-Based Dynamic Prediction Model}
\usage{
plot_dynamicHR(object,variable,form_splines,IDnames=NULL,xlab = "Follow-up Time (years)",ylab = "Dynamic Hazard Ratio",ylim =NULL, col = c("black", "black", "black"), lty = c(1, 2, 2),lwd=6,...)
}
\arguments{
\item{object}{An objects inheriting from class "jm".}

\item{variable}{Name of longitudinal covariates.}

\item{form_splines}{Form of the dynamic effects function.It mainly includes spline functions (spline) and quadratic functions (quadratic).}

\item{tt}{Follow-up time with the maximum of interest.}

\item{IDnames}{Label of the plot, with the default being NULL;}

\item{legend}{Annotations or descriptions of the plot, with the default being NULL;}

\item{xlab}{Label of the horizontal axis, with the default being "Follow-up Time (years)".}

\item{ylab}{Label of the vertical axis, with the default being "Dynamic Hazard Ratio".}

\item{ylim}{The scale range of the vertical axis is, by default, from the minimum to the maximum value of effect size of the covariate.}

\item{col}{Line Color Options, with the default being "black".}

\item{lty}{Line Type Options.By default, the HR is represented by a solid line, and the confidence interval is represented by a dashed line.}

\item{lwd}{Line Thickness Options , with the default being "lwd=6".}

\item{...}{extra aguments; currently none is used.}
}
\description{
Title Dynamic Effects Plot of Covariates in a Joint Model-Based Dynamic Prediction Model
}
\examples{
library(JMbayes2)
library(DynamicModel)
data(renal)
renal.id<-renal[!duplicated(renal$id),]
fit1<-lme(GFR~age+weight+ns(yearse,3), random =~yearse | id,data = renal,control = list(opt = "optim"))
fit2<-lme(proteinuria~age+ns(yearse,3), random = ~yearse | id,  data =renal,control = list(opt = "optim"))
cox<-coxph(Surv(time,status)~age+sex+weight,data=renal.id,x=TRUE)
form_splines <- list("GFR"=~ value(GFR)* (yearse +I(yearse^2)),
"proteinuria"=~ value(proteinuria) * (yearse+I(yearse^2)))
jointFit1 <- jm(cox, list(fit1,fit2), time_var = "yearse",seed=1,n_chains = 1L, functional_forms = form_splines)
plot_dynamicHR(jointFit1,"GFR",form_splines="quadratic",tt=15,IDnames="b.",ylim=c(0,0.5),xlim=c(0,15),cex.axis=1.5,cex.lab=1.5)
}
