% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_LM.R
\name{predict_LM}
\alias{predict_LM}
\title{Title Individual dynamic prediction function.}
\usage{
predict_LM(object,newdata)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM"}

\item{newdata}{data.frame}
}
\value{
\item{out} Predicted values.
}
\description{
Title Individual dynamic prediction function.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
newdata<-renal[which(renal$id==5619),]
aa<-predict_LM(fit,newdata)
}
