% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_LM.R
\name{predictive_LM}
\alias{predictive_LM}
\title{Predictive Performance Assessment for Landmarking Models}
\usage{
predictive_LM(object,method,n,K,seed,r)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM"}

\item{method}{Character string specifying validation method: "K-fold", "boot", or "split"}

\item{n}{Number of repetitions for internal cross-validation. For random splitting, it refers to the number of random splitting iterations; for K-fold cross-validation, it refers to the number of outer loop iterations; for bootstrap, it refers to the number of resampling iterations.}

\item{K}{Integer, number of folds for K-fold cross-validation (only used when method="K-fold").}

\item{seed}{Integer, random seed for reproducibility}

\item{r}{Numeric, proportion of data to use for training (only used when method="split")}

\item{n_group}{Integer, number of groups for calibration (default=10)}

\item{dca_thresholds}{Numeric vector specifying the threshold probabilities used for decision curve analysis (DCA). The default is {seq(0.05, 0.50, by = 0.01)}}
}
\value{
A list containing:
\itemize{
  \item{time: The sequence of landmark times used for evaluation}
  \item{cindex_corr: Corrected concordance index with confidence intervals}
  \item{BS_corr: Corrected Brier score with confidence intervals}
  \item{AUC_corr: Corrected AUC with confidence intervals}
  \item{calibration_slpoe_with_ci: Calibration slope with confidence intervals}
  \item{all_cal_autual: Complete calibration results}
  \item{dca_results: Decision curve analysis results, including threshold probabilities and net benefit for the prediction model, treat-all strategy, and treat-none strategy}
}
}
\description{
This function evaluates the predictive performance of landmarking models using
different validation methods including K-fold cross-validation, bootstrap validation,
and data splitting.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal ,tw=list(sl=seq(0,10, by=1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
a1<-VS_LM(fit,cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")))
LM_pre<-predictive_LM(a1,method="K-fold",n=1,seed=1,K=5)
}
