% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LMf.R
\name{print.LMf}
\alias{print.LMf}
\title{Title output result of function "LMf"}
\usage{
\method{print}{LMf}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class LMf}
}
\description{
Title output result of function "LMf"
}
