% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VS_LM.R
\name{print.VS_LM}
\alias{print.VS_LM}
\title{Title output result of function "VS_LM"}
\usage{
\method{print}{VS_LM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class VS_LM}
}
\description{
Title output result of function "VS_LM"
}
