% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal.R
\name{renal}
\alias{renal}
\title{Renal transplantation data}
\format{
A numeric data.frame
\describe{
  \item{id}{Patients identifier; in total there are 407 patients.}
  \item{time}{Time-to-event data for renal graft failure or maximum follow up time, with transplant date as the time origin (years).}
  \item{status}{Censoring indicator (1 = graft failure and 0 = censored).}
  \item{age}{Column of patient's age divided by 10.}
  \item{weight}{Column of patient's weight divided by 10.}
  \item{sex}{Gender of patient.}
  \item{hema}{Hematocrit, recorded as percentage (0.1\%) of the ratio of the volume of red blood cells to the total volume of blood.}
  \item{GFR}{Measured as 10 ml/min/1.73 m².}
  \item{proteinuria}{Measured as percentage 1 g/24 hours.}
  \item{yearse}{Observation time.}
}
 @export
 @references
 \describe{
 \item{Rizopoulos D, Ghosh, P. A Bayesian semiparametric multivariate joint model for multiple longitudinal outcomes and a time-to-event. Stat Med. 2011; 30(12): 1366-80.}}
 @examples
 data(renal)
 head(renal)
}
\usage{
data(renal)
}
\description{
This is a dataset on 407 patients suffering from chronic kidney disease who underwent a primary renal transplantation with a graft from a deceased or living donor in the University Hospital of the Catholic University of Leuven (Belgium) between 21 January 1983 and 16 August 2000.
}
