% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMf.R
\name{LMf}
\alias{LMf}
\title{Title The Construction of dynamic prediction model based on landmark}
\usage{
LMf (data,tw=list(sl=sl,w=w),cov=list(fixed=fixed,vary=vary),id=id,rtime=rtime,time=time,status=status,inter=T,func_covars=c("linear","quadratic"),func_lms=c("linear","quadratic"))
}
\arguments{
\item{data}{data.frames.}

\item{tw}{A list of time point options, including landmark time points (sl) and prediction windows (w).}

\item{cov}{A list of covariates selection options, including baseline covariates: fixed and longitudinal covariates: vary.}

\item{id}{Individual ID}

\item{rtime}{Observation time column name}

\item{time}{Individual's survival time column name}

\item{status}{Individual's survival status column name}

\item{pre_cov}{Specify covariates with dynamic effects in advance}

\item{inter}{Indicator variables: whether covariates in the model need to include dynamic effects, with the default being TRUE.}

\item{func_covars}{The functional form of the dynamic effect of covariates. When inter is TRUE, the available functional forms include linear functions (linear), quadratic functions (quadratic), exponential functions (exp), and logarithmic functions (log), with the default being linear and quadratic functions.}

\item{func_lms}{The functional forms for the dynamic effects of baseline hazard rates. When inter is TRUE, the available functional forms include linear functions (linear), quadratic functions (quadratic), exponential functions (exp), and logarithmic functions (log), with the default being linear and quadratic functions.}
}
\value{
A list of class LMf with components:
 \describe{
   \item{Model}{The result of parameter estimation.}
   \item{data}{The Landmark dataset.}
   \item{func_covars}{Functional forms for the dynamic effects of covariates.}
   \item{func_lms}{Functional forms for the dynamic effects of baseline hazard rates.}
   \item{cov}{A list of covariates type. The components of this list are:
     \itemize{
       \item{fixed}{Baseline covariates.}
       \item{vary}{Longitudinal covariates.}
     }
   }
   \item{tw}{A list of time points. The components of this list are:
     \itemize{
       \item{sl}{Landmark time points.}
       \item{w}{Prediction window.}
     }
   }
   \item{time}{Individual's survival time.}
   \item{status}{Individual's survival status.}
   \item{id}{Individual ID.}
   \item{rtime}{Observation time column name.}
 }
}
\description{
Title The Construction of dynamic prediction model based on landmark
}
\examples{
library(DynamicModel)
data(renal)
head(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
print(fit)
}
\references{
Fries AH, Choi E, Wu JT, Lee JH, Ding VY, Huang RJ, et al. Software Application Profile: dynamicLM —a tool for performing dynamic risk prediction using a landmark supermodel for survival data under competing risks. International Journal of Epidemiology 2023;52:1984–9. https://doi.org/10.1093/ije/dyad122.

Yang Z, Hou Y, Lyu J, Liu D, Chen Z. Dynamic prediction and prognostic analysis of patients with cervical cancer: a landmarking analysis approach. Annals of Epidemiology 2020;44:45–51. https://doi.org/10.1016/j.annepidem.2020.01.009.
}
