% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cHR.R
\name{plot_cHR}
\alias{plot_cHR}
\title{Title Function for dynamic effect hazard ratio (HR) plots.}
\usage{
plot_cHR(object,covars,conf_int=TRUE,IDlegend=NULL,legend=NULL,xlab="landmark time points",ylab="Dynamic HR",ylim,...)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM"}

\item{covars}{Covariate names.}

\item{conf_int}{Indicator variable. It indicates whether confidence intervals are included.}

\item{IDlegend}{Annotations or descriptions of the plot.}

\item{legend}{Label of the plot.}

\item{xlab}{Label of the horizontal axis}

\item{ylab}{Label of the vertical axis.}

\item{ylim}{Scale of the vertical axis}

\item{...}{extra aguments; currently none is used.}
}
\description{
Title Function for dynamic effect hazard ratio (HR) plots.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=T)
plot_CHR(fit,c("age"),conf_int=TRUE,IDlegend="a.",legend="(Per 10 years)",type="Cox",xlab="landmark time points(s)",ylab="Dynamic HR (w=5)")
}
