% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamicHR.R
\name{plot_dynamicHR}
\alias{plot_dynamicHR}
\title{#' This function creates plots of dynamic hazard ratios over time from joint models,
showing the time-varying effect of covariates on the hazard function.}
\usage{
plot_dynamicHR(object,variable,form_splines)
}
\arguments{
\item{object}{An objects inheriting from class "jm".}

\item{variable}{Character string specifying the covariate for which to plot the hazard ratio}

\item{form_splines}{Character string specifying the functional form: "quadratic" or "spline"}

\item{tt}{Numeric, the maximum follow-up time for the plot}

\item{IDnames}{Character string for patient identifier to display (optional, default=NULL)}

\item{legend}{Character vector for legend labels (optional, default=NULL)}

\item{xlab}{Character string for x-axis label (default="Follow-up Time (years)")}

\item{ylab}{Character string for y-axis label (default="Dynamic Hazard Ratio")}

\item{ylim}{Numeric vector of length 2 specifying y-axis limits (optional)}

\item{col}{Vector of colors for the mean, lower and upper bounds (default=c("black", "black", "black"))}

\item{lty}{Vector of line types for the mean, lower and upper bounds (default=c(1, 2, 2))}

\item{lwd}{Line width for the plot (default=6)}

\item{axis_lwd}{Line width for axes (default=1.6)}

\item{axis_tcl}{Tick length for axes (default=-0.4)}

\item{axis_padj_x}{Vertical adjustment for x-axis labels (default=-0.3)}

\item{axis_hadj_y}{Horizontal adjustment for y-axis labels (default=0.9)}

\item{axis_las}{Orientation of axis labels (0=parallel, 1=horizontal, 2=perpendicular, 3=vertical) (default=1)}

\item{title_font_lab}{Font for axis labels (default=7 for bold)}

\item{title_line_xlab}{Line position for x-axis label (default=2.6)}

\item{title_line_ylab}{Line position for y-axis label (default=2.6)}

\item{main_title_font}{Font for main title (default=7 for bold)}

\item{main_title_cex}{Character expansion for main title (default=1.5)}

\item{main_title_line}{Line position for main title (default=1)}

\item{abline_lty}{Line type for reference line at HR=1 (default=2 for dashed)}

\item{idnames_side}{Side to display ID names (1=bottom, 2=left, 3=top, 4=right) (default=3)}

\item{idnames_line}{Line position for ID names (default=1)}

\item{idnames_adj}{Adjustment for ID names (0=left, 1=right) (default=0)}

\item{idnames_cex}{Character expansion for ID names (default=1.5)}

\item{idnames_font}{Font for ID names (default=7 for bold)}

\item{legend_pos}{Position of legend (default="topright")}

\item{legend_lty}{Line type in legend (default=1)}

\item{legend_lwd}{Line width in legend (default=6)}

\item{legend_bty}{Box type for legend (default="n" for no box)}

\item{mar}{Numeric vector of length 4 specifying margins (bottom, left, top, right) (default=c(5, 5, 4, 2))}

\item{...}{extra aguments; currently none is used.}
}
\description{
#' This function creates plots of dynamic hazard ratios over time from joint models,
showing the time-varying effect of covariates on the hazard function.
}
\examples{
library(JMbayes2)
library(DynamicModel)
data(renal)
renal.id<-renal[!duplicated(renal$id),]
fit1<-lme(GFR~age+weight+ns(yearse,3), random =~yearse | id,data = renal,control = list(opt = "optim"))
fit2<-lme(proteinuria~age+ns(yearse,3), random = ~yearse | id,  data =renal,control = list(opt = "optim"))
cox<-coxph(Surv(time,status)~age+sex+weight,data=renal.id,x=TRUE)
form_splines <- list("GFR"=~ value(GFR)* (yearse +I(yearse^2)),
"proteinuria"=~ value(proteinuria) * (yearse+I(yearse^2)))
jointFit1 <- jm(cox, list(fit1,fit2), time_var = "yearse",seed=1,n_chains = 1L, functional_forms = form_splines)
plot_dynamicHR(jointFit1,"GFR",form_splines="quadratic",tt=15,IDnames="b.",ylim=c(0,0.5),xlim=c(0,15),cex.axis=1.5,cex.lab=1.5)
}
