% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict.R
\name{plot_predict}
\alias{plot_predict}
\title{Title Function for individual dynamic prediction plots.}
\usage{
plot_predict(object,newdata,ylim,main="Patient",xlab="landmark time points (s)",ylab="5-year survival",legend=NULL,...)
}
\arguments{
\item{object}{An objects inheriting from class "LMf" or "Vs_LM".}

\item{newdata}{The data.frame.}

\item{ylim}{The scale range of the vertical axis is, by default, from 0 to 1.}

\item{main}{Title of the plot, with default being "Patient".}

\item{xlab}{Label of the horizontal axis, with default being "landmark time points (s)".}

\item{ylab}{Label of the vertical axis, with default being "5-year survival".}

\item{legend}{Label of the plot, with default NULL.}

\item{...}{extra aguments; currently none is used.}
}
\description{
Title Function for individual dynamic prediction plots.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
a1<-VS_LM(fit,cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")))
newdata<-renal[which(renal$id==5619),]
plot_predict(object=a1,newdata=newdata,main="")
}
