% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_LM.R
\name{predict_LM}
\alias{predict_LM}
\title{Predict Survival Probabilities from Landmarking Models}
\usage{
predict_LM(object,newdata)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM"}

\item{newdata}{A data.frame containing the new data for which predictions are made}
}
\value{
A data.frame with two columns:
\itemize{
  \item{time: The landmark time points}
  \item{csurv: The conditional survival probabilities from each landmark time point
               to the landmark time plus the prediction window}
}
}
\description{
This function generates survival probability predictions from a fitted landmarking model
at specified landmark time points for new data.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
a1<-VS_LM(fit,cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")))
newdata<-renal[which(renal$id==5619),]
aa<-predict_LM(a1,newdata)
}
