% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_JM.R
\name{predictive_JM}
\alias{predictive_JM}
\title{Title}
\usage{
predictive_JM(jointFit,data,method,formulas,random_effect_formulas,form_splines,surv_formula,seq_len,w,n,K,seed,r)
}
\arguments{
\item{jointFit}{an objects inheriting from class "jm"}

\item{data}{Dataset for fitting the joint model (JM).}

\item{method}{Types of internal cross validation. There are three types of internal validation methods: random splitting:sample, K-fold cross-validation:CV, and bootstrap:boot.}

\item{formulas}{A list of functional forms for fixed effects.}

\item{random_effect_formulas}{A list of functional forms for random effects.}

\item{form_splines}{A list of functional forms for the dynamic effects of longitudinal covariates.}

\item{surv_formula}{Functional form of the survival submodel.}

\item{seq_len}{a numeric vector of future times to calculate predictions}

\item{w}{prediction window}

\item{n}{umber of repetitions for internal cross-validation}

\item{K}{The specific number of folds in the K-fold cross-validation method.}

\item{seed}{The seed number for sampling.}

\item{r}{The proportion for random splitting.}
}
\value{
A list of class LMf with components:
\describe{
\item {sl} landmark time points
\item {cindex_corr} C-index
\item {BS_corr} Brier score
\item {AUC_corr} AUC
}
}
\description{
Title
}
\examples{
library(JMbayes2)
library(DynamicModel)
data(renal)
renal.id<-renal[!duplicated(renal$id),]
fit1<-lme(GFR~age+weight+ns(yearse,3), random =~yearse | id,data = renal,control = list(opt = "optim"))
fit2<-lme(proteinuria~age+ns(yearse,3), random = ~yearse | id,  data =renal,control = list(opt = "optim"))
cox<-coxph(Surv(time,status)~age+sex+weight,data=renal.id,x=TRUE)
form_splines <- list("GFR"=~ value(GFR)* (yearse +I(yearse^2)),
"proteinuria"=~ value(proteinuria) * (yearse+I(yearse^2)))
jointFit1 <- jm(cox, list(fit1,fit2), time_var = "yearse",seed=1,n_chains = 1L, functional_forms = form_splines)
aa<-predictive_JM(jointFit1,renal,method="CV",formulas=list(GFR~weight+age+ns(yearse,3),proteinuria~age+ns(yearse,3)),random_effect_formulas=list(~yearse|id,~yearse|id),surv_formula=Surv(time,status)~age+sex+weight,form_splines=form_splines,seq_len=seq(0,10,by=1),w=5,n=1,K=5,seed=1)
}
