% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_JM_ex.R
\name{predictive_JM_ex}
\alias{predictive_JM_ex}
\title{Title External Validation Function for Joint model-Based Dynamic Prediction Models}
\usage{
predictive_JM_ex(object, newdata, seq_len, w)
}
\arguments{
\item{object}{an objects inheriting from class "jm"}

\item{newdata}{An external validation dataset with variable names identical to those in the model-building dataset.}

\item{seq_len}{a numeric vector of future times to calculate predictions}

\item{w}{prediction window}
}
\value{
A list of class predictive_JM with components:
\describe{
\item {time} landmark time points
\item {cindex} C-index
\item {BS} Brier score
\item {AUC} AUC
}
}
\description{
Title External Validation Function for Joint model-Based Dynamic Prediction Models
}
