% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_LM.R
\name{predictive_LM}
\alias{predictive_LM}
\title{Title Internal cross-validation function for landmark-based dynamic prediction models.}
\usage{
predictive_LM(object,method,n,K,seed,r)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM"}

\item{method}{Types of internal cross validation.There are three types of internal validation methods: random splitting:sample, K-fold cross-validation:CV, and bootstrap:boot.}

\item{n}{Number of repetitions for internal cross-validation}

\item{K}{The specific number of folds in the K-fold cross-validation method.}

\item{seed}{The seed number for sampling.}

\item{r}{The proportion for random splitting.}
}
\value{
A list of class LMf with components:
\describe{
\item {sl} landmark time points
\item {cindex_corr} C-index
\item {BS_corr} Brier score
\item {AUC_corr} AUC
}
}
\description{
Title Internal cross-validation function for landmark-based dynamic prediction models.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=T)
LM_pre<-predictive_LM(fit,method="CV",n=1,seed=1,K=5)
}
