% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_LM_ex.R
\name{Predictive_LM_ex}
\alias{Predictive_LM_ex}
\title{Title External Validation Function for Landmarking-Based Dynamic Prediction Models}
\usage{
Predictive_LM_ex(object, ex_data)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM".}

\item{ex_data}{An external validation dataset with variable names identical to those in the model-building dataset.}
}
\value{
A list of class Predictive_LM_ex with components:
\describe{
\item {time} landmark time points
\item {cindex_corr} C-index
\item {BS_corr} Brier score
\item {AUC_corr} AUC
}
}
\description{
Title External Validation Function for Landmarking-Based Dynamic Prediction Models
}
