% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_LM_ex.R
\name{Predictive_LM_ex}
\alias{Predictive_LM_ex}
\title{External Validation for Landmarking Models}
\usage{
Predictive_LM_ex(object, ex_data, n_group = 10)
}
\arguments{
\item{object}{an objects inheriting from class "LMf" or "Vs_LM".}

\item{ex_data}{An external validation dataset with variable names identical to those in the model-building dataset.}

\item{n_group}{Integer, number of groups for calibration (default=10)}
}
\value{
A list containing:
\itemize{
  \item{time: The sequence of landmark times used for evaluation}
  \item{cindex: Concordance index values for each landmark time}
  \item{BS: Brier score values for each landmark time}
  \item{AUC: AUC values for each landmark time}
  \item{calibration_slope_with_ci: Calibration slope with confidence intervals}
  \item{all_cal_autual: Complete calibration results}
}
}
\description{
This function performs external validation of landmarking models on an external dataset.
It evaluates discrimination (C-index, AUC), calibration, and overall performance (Brier score)
at multiple landmark times.
}
