% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal.R
\name{renal}
\alias{renal}
\title{Renal graft failure dataset}
\format{
A data.frame
\describe{
  \item{id}{Patients identifier; in total there are 407 patients.}
  \item{time}{Time-to-event data for renal graft failure or maximum follow up time, with transplant date as the time origin (years).}
  \item{status}{Censoring indicator (1 = graft failure and 0 = censored).}
  \item{age}{Column of patient's age divided by 10.}
  \item{weight}{Column of patient's weight divided by 10.}
  \item{sex}{Gender of patient.}
  \item{hema}{Hematocrit, recorded as percentage (0.1\%) of the ratio of the volume of red blood cells to the total volume of blood.}
  \item{GFR}{Measured as 10 ml/min/1.73 m².}
  \item{proteinuria}{Measured as percentage 1 g/24 hours.}
  \item{yearse}{Observation time.}
}
}
\usage{
data(renal)
}
\description{
This is a dataset on 407 patients suffering from chronic kidney disease who underwent a primary renal transplantation with a graft from a deceased or living donor in the University Hospital of the Catholic University of Leuven (Belgium) between 21 January 1983 and 16 August 2000.
}
\examples{
# Load the dataset
data(renal)

# View the first few rows of the dataset
head(renal)
}
\references{
Rizopoulos D, Ghosh P. A Bayesian semiparametric multivariate joint model for multiple longitudinal outcomes and a time‐to‐event. Statistics in Medicine 2011;30:1366–80. https://doi.org/10.1002/sim.4205.
}
