% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS_LM.R
\name{VS_LM}
\alias{VS_LM}
\title{Title Function for variable selection in landmark-based dynamic prediction models.}
\usage{
VS_LM(object,cov=list(fixed=fixed,vary=vary))
}
\arguments{
\item{object}{an objects inheriting from class "LMf"}

\item{cov}{A list of covariates selection options, including baseline covariates: fixed and longitudinal covariates: vary.}
}
\value{
A list of class LMf with components:
 \describe{
   \item{Model}{The result of parameter estimation.}
   \item{data}{The Landmark dataset.}
   \item{func_covars}{Functional forms for the dynamic effects of covariates.}
   \item{func_lms}{Functional forms for the dynamic effects of baseline hazard rates.}
   \item{cov}{A list of covariates type. The components of this list are:
     \itemize{
       \item{fixed}{Baseline covariates.}
       \item{vary}{Longitudinal covariates.}
     }
   }
   \item{tw}{A list of time points. The components of this list are:
     \itemize{
       \item{sl}{Landmark time points.}
       \item{w}{Prediction window.}
     }
   }
   \item{time}{Individual's survival time.}
   \item{status}{Individual's survival status.}
   \item{id}{Individual ID.}
   \item{rtime}{Observation time column name.}
 }
}
\description{
Title Function for variable selection in landmark-based dynamic prediction models.
}
\examples{
library(DynamicModel)
data(renal)
fit<-LMf(renal,tw=list(sl=seq(0, 10, by=0.1),w=5),cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")),id="id",rtime="yearse",time="time",status="status",inter=TRUE)
a1<-VS_LM(fit,cov=list(fixed=c("age", "weight", "sex"),vary=c("GFR","proteinuria")))
print(a1)
}
